+++ src/client/src/ui/spotTile/components/notifications/TileRejected.tsx
